requirements.txt
streamlit

langchain==0.2.16
langchain-community==0.2.16
langchain-groq

python-dotenv
bs4
requests

langchain-text-splitters

ddgs

pypdf

sentence-transformers==2.7.0
torch

chromadb==0.4.24

pandas
numpy
scikit-learn